requirements.txt
selenium==4.12.0
undetected-chromedriver==3.5.3
python-dotenv==1.0.1